+++ src/main/java/kr/hhplus/be/server/domain/coupon/service/CouponService.java
@@ -76,7 +76,7 @@ public class CouponService {
     }
 
     // 사용자 쿠폰 목록 조회
-    public List<UserCouponResponse> getUserCoupon(int userId){
+    public List<UserCouponResponse> getUserCoupon(long userId){
         List<UserCoupon> userCouponList = couponRepository.getUserCoupon(userId);
 
         if(userCouponList.isEmpty()){
