+++ src/main/java/kr/hhplus/be/server/interfaces/coupon/dto/UserCouponResponse.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UserCouponResponse {
-    private int userId;
-    private int couponId;
+    private long userId;
+    private long couponId;
 }
