+++ src/main/java/kr/hhplus/be/server/interfaces/point/dto/UserPoint.java
@@ -1,11 +1,11 @@
 package kr.hhplus.be.server.interfaces.point.dto;
 
 public record UserPoint(
-        int id,
+        long id,
         int point,
         long updateMillis
 ) {
-    public static UserPoint empty(int id) {
+    public static UserPoint empty(long id) {
         return new UserPoint(id, 0, System.currentTimeMillis());
     }
 }
