+++ src/test/java/kr/hhplus/be/server/domain/point/service/PointServiceTest.java
@@ -48,7 +48,7 @@ class PointServiceTest {
     @DisplayName("사용자의 포인트 충전")
     void 포인트_충전(){
         //given
-        int userId = 1;
+        long userId = 1L;
         int amount = 10000;
 
         Point point = Point.builder()
@@ -72,7 +72,7 @@ class PointServiceTest {
     @DisplayName("사용자의 포인트 충전 실패 케이스: 10000단위가 아닐시")
     void 포인트_충전2(){
         //given
-        int userId = 1;
+        long userId = 1L;
         int amount = 10001;
 
         Point point = Point.builder()
@@ -95,7 +95,7 @@ class PointServiceTest {
     @DisplayName("사용자의 포인트 충전 실패 케이스: 백만 포인트 이상 충전시")
     void 포인트_충전3(){
         //given
-        int userId = 1;
+        long userId = 1L;
         int amount = 1000000;
 
         Point point = Point.builder()
@@ -118,7 +118,7 @@ class PointServiceTest {
     @DisplayName("사용자의 포인트 충전 실패 케이스: 0포인트이하 충전시")
     void 포인트_충전4(){
         //given
-        int userId = 1;
+        long userId = 1L;
         int amount = 0;
 
         Point point = Point.builder()
@@ -142,7 +142,7 @@ class PointServiceTest {
     @DisplayName("사용자의 포인트 충전 실패 케이스: 충전시 누적 포인트 백만 포인트 초과하면 안됨")
     void 포인트_충전5(){
         //given
-        int userId = 1;
+        long userId = 1L;
         int amount = 600000;//충전할양
 
         Point point = Point.builder()
